requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
openpyxl==3.1.2

# Mini app
Flask==3.0.3
requests==2.32.3

